backend/pulse-ingestion-service/requirements.txt
confluent-kafka[avro]==2.4.0
fastavro==1.9.5
python-dotenv==1.0.1
praw==7.7.1
redis==5.0.1